Extract platform capability flag normalization
