+++ .github/workflows/quality-baseline.yml
@@ -47,7 +47,7 @@ jobs:
         run: |
           mkdir -p output/quality-baseline
           python scripts/check_refactor_quality_thresholds.py \
-            --max-source-file-lines 404 \
+            --max-source-file-lines 402 \
             --max-function-lines 49 \
             2>&1 | tee output/quality-baseline/refactor-thresholds.txt
