+++ quality/architecture_rules.md
@@ -35,11 +35,14 @@ report-only baseline is reviewed.
 
 The largest current modularity risks are:
 
-1. `src/app/services/platform_capabilities_normalization.py` at 404 lines,
-2. `src/app/services/performance_workspace_contribution.py` at 402 lines,
-3. `src/app/services/risk_workspace_service.py` at 402 lines,
-4. `src/app/services/dpm_proof_pack_service.py` at 399 lines,
-5. `src/app/contracts/advisor_brief.py` at 398 lines.
+1. `src/app/services/performance_workspace_contribution.py` at 402 lines,
+2. `src/app/services/risk_workspace_service.py` at 402 lines,
+3. `src/app/services/dpm_proof_pack_service.py` at 399 lines,
+4. `src/app/contracts/advisor_brief.py` at 398 lines.
+
+`src/app/services/platform_capabilities_normalization.py` is reduced from 404 to 192 lines after
+source capability feature-key and workflow-key interpretation moved into
+`src/app/services/platform_capabilities_feature_flags.py`.
 
 `src/app/contracts/proposal_lifecycle.py` is reduced from 405 to 21 lines after summary,
 workflow, lineage, and envelope DTOs moved into focused lifecycle contract modules.
